+++ acq4/analysis/modules/pbm_ImageAnalysis/pbm_ImageAnalysis.py
@@ -600,7 +600,7 @@ class pbm_ImageAnalysis(AnalysisModule):
         self.specImageCalcFlag = False  # we need to recalculate the spectrum
         npts = self.imageData.shape[0]/2
         freq = np.fft.fftfreq(npts, d=self.imagedT)
-        freq = freq[0:npts/2 + 1]
+        freq = freq[0:int(npts/2 + 1)]
         self.ctrlROIFunc.ImagePhys_SpecHPF.setMinimum(0.0)
         self.ctrlROIFunc.ImagePhys_SpecHPF.setMaximum(np.max(freq))
         self.ctrlROIFunc.ImagePhys_SpecHPF.setValue(freq[1])
@@ -748,7 +748,7 @@ class pbm_ImageAnalysis(AnalysisModule):
             
         npts = self.imageData.shape[0]/2
         freq = np.fft.fftfreq(npts, d=self.imagedT)  # get frequency list
-        freq = freq[0:npts/2 + 1]
+        freq = freq[0:int(npts/2 + 1)]
         hpf = self.ctrlROIFunc.ImagePhys_SpecHPF.value()
         lpf = self.ctrlROIFunc.ImagePhys_SpecLPF.value()
         u = np.where(freq > hpf) # from frequencies, select those from the window
@@ -2564,7 +2564,7 @@ class pbm_ImageAnalysis(AnalysisModule):
             Fy = np.fft.fft(yanom, npad, axis=axis)
             iFxy = np.fft.ifft(Fx.conj()*Fy,n=npad,axis=axis).real
         # We juste turn the lags into correct positions:
-        iFxy = np.concatenate((iFxy[len(iFxy)/2:len(iFxy)],iFxy[0:len(iFxy)/2]))
+        iFxy = np.concatenate((iFxy[int(len(iFxy)/2):len(iFxy)],iFxy[0:int(len(iFxy)/2)]))
         return iFxy/varxy
 
 #
