+++ acq4/modules/NoiseMonitor.py
@@ -254,7 +254,7 @@ class ChannelRecorder(Qt.QSplitter):
 
         # Spectrum analysis
         fft = np.abs(np.fft.fft(dataArr))
-        fft = fft[:len(fft)/2]
+        fft = fft[:int(len(fft)/2)]
         freqArr = np.linspace(0, rate/2., len(fft))
 
         # downsample spectrogram
